Move method (again) next to tightly coupled method
proposePeers is only used by proposePeersFromRepository and is a
continuation of its logic, so moving next to it.
